+++ Model/Model.h
@@ -11,7 +11,7 @@ class Model {
 public:
     [[nodiscard]] int getParticleCount() const;
     void getParticlePosition(int index, float &x, float &y) const;
-    [[nodiscard]] int getParticleRadius() const;
+    [[nodiscard]] float getParticleRadius() const;
 private:
     World world_;
 };
